g4-common: Prebuilt android.hardware.camera.flash-autofocus.xml
It all started when i used prebuilt power.default.so and
camera.msm8992.so to achieve max camera quality, equal to LG stock
roms.
This broke the torch tile because the LG hals are not compatible with
the Oreo and now Pie framework flashlight controller.
So in our case, we wanted to keep the stock hals because we needed
the maximum possible camera quality and shot amazing photos and
we created an alternative method to get the torch tile working.
'commit: 0b37f02b269232c2de6002523fe928e4b1c6ffb8'

Remove now the torch tile created by the framework flashlight controller
to keep anyone away from pressing it and crashing the cameraserver

Change-Id: I924780f60c0950328d4f0ccca72f9ed5d5ac2440