g4-common: init: Write the pid of services
Write the pid of all the services and only of those we need
to always run in the background, to system background tasks.

{ Also 'disabled' flag should only be used on services which
we want to disable or to services which we want to add another
flag and start them automatically at a specific time. Clean up
this mess with all those 'disabled' flags }

Change-Id: I630a0cba47b2b5fd42cd6c80e1b1746cde7bfea0